[DEV-5291] cleanup
